+++ libraries/extractor/src/main/java/androidx/media3/extractor/text/SubtitleTranscodingTrackOutput.java
@@ -26,6 +26,7 @@ import androidx.media3.common.C;
 import androidx.media3.common.DataReader;
 import androidx.media3.common.Format;
 import androidx.media3.common.MimeTypes;
+import androidx.media3.common.util.Log;
 import androidx.media3.common.util.ParsableByteArray;
 import androidx.media3.common.util.Util;
 import androidx.media3.extractor.TrackOutput;
@@ -40,6 +41,8 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
  */
 /* package */ final class SubtitleTranscodingTrackOutput implements TrackOutput {
 
+  private static final String TAG = "SubtitleTranscodingTO";
+
   private final TrackOutput delegate;
   private final SubtitleParser.Factory subtitleParserFactory;
   private final CueEncoder cueEncoder;
@@ -50,6 +53,7 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
   private byte[] sampleData;
   @Nullable private SubtitleParser currentSubtitleParser;
   private @MonotonicNonNull Format currentFormat;
+  private boolean shouldSuppressParsingErrors;
 
   public SubtitleTranscodingTrackOutput(
       TrackOutput delegate, SubtitleParser.Factory subtitleParserFactory) {
@@ -66,6 +70,16 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
     if (currentSubtitleParser != null) {
       currentSubtitleParser.reset();
     }
+  }
+
+  /**
+   * Sets whether to suppress parsing errors thrown during the transcoding in {@link
+   * #sampleMetadata}.
+   *
+   * <p>Defaults to {@code false}.
+   */
+  public void shouldSuppressParsingErrors(boolean shouldSuppressParsingErrors) {
+    this.shouldSuppressParsingErrors = shouldSuppressParsingErrors;
   }
 
   // TrackOutput implementation
@@ -143,12 +157,21 @@ import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
     checkArgument(cryptoData == null, "DRM on subtitles is not supported");
 
     int sampleStart = sampleDataEnd - offset - size;
-    currentSubtitleParser.parse(
-        sampleData,
-        sampleStart,
-        size,
-        SubtitleParser.OutputOptions.allCues(),
-        cuesWithTiming -> outputSample(cuesWithTiming, timeUs, flags));
+    try {
+      currentSubtitleParser.parse(
+          sampleData,
+          sampleStart,
+          size,
+          SubtitleParser.OutputOptions.allCues(),
+          cuesWithTiming -> outputSample(cuesWithTiming, timeUs, flags));
+    } catch (RuntimeException e) {
+      if (shouldSuppressParsingErrors) {
+        // TODO: b/391362063 - Propagate this error out in a non-fatal way.
+        Log.w(TAG, "Parsing subtitles failed, ignoring sample.", e);
+      } else {
+        throw e;
+      }
+    }
     sampleDataStart = sampleStart + size;
     if (sampleDataStart == sampleDataEnd) {
       // The array is now empty, so we can move the start and end pointers back to the start.